Add additional measurements and export datasets
This adds graphs with moving averages for each wiki in order to start
discussing whether that is useful information to have. Also adds
a modification to the "non-reverted non-reblocked expired block"
where the non-reverted edit has to be in the article namespace
(namespace 0, also called "Main").

Also adds code to export all the aggregated datasets so that they
can be shared and discussed within the team.