Add partial block measurements for more wikis with graphs
Per discussion within the team, adding Arabic, Persian, Meta,
and Mediawiki to the wikis with partial blocks. The data gathering
and analysis of partial blocks is modified to iterate over a list
of wikis, thus enabling later expansion as necessary. Graphs
showing the number of blocks per day, with one- and two-week
moving averages are added as well.

Committing updates to several other graphs as well.